requirements.txt
GROQ_API_KEY 
NEWS_API_KEY
streamlit
langchain
